% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{kormap1}
\alias{kormap1}
\title{A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area(2010), level 1}
\format{A Shape data with length of 16}
\usage{
kormap1
}
\description{
A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area(2010), level 1
}
\examples{
qtm("kormap1")
}
\keyword{datasets}

