% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{kormap2}
\alias{kormap2}
\title{A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area(2010), level 2}
\format{A Shape data with length of 251}
\usage{
kormap2
}
\description{
A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area(2010), level 2
}
\keyword{datasets}

