% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{kormap3}
\alias{kormap3}
\title{A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area(2010), level 3}
\format{A Shape data with length of 3472}
\usage{
kormap3
}
\description{
A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area(2010), level 3
}
\keyword{datasets}

