% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{korpopmap1}
\alias{korpopmap1}
\title{A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area, level 1
with merged population census data 2010}
\format{A Shape data with length of 16}
\usage{
korpopmap1
}
\description{
A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area, level 1
with merged population census data 2010
}
\keyword{datasets}

