% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{korpopmap2}
\alias{korpopmap2}
\title{A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area, level 2
with merged population census data 2010}
\format{A Shape data with length of 251}
\usage{
korpopmap2
}
\description{
A data of SpatialPolygonsDataFrame contains the map of  Korean administrative area, level 2
with merged population census data 2010
}
\keyword{datasets}

