% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myleaflet.R
\name{myleaflet}
\alias{myleaflet}
\title{Create a Leaflet map widget with multiple map layers and control}
\usage{
myleaflet(map, choices, popupcol = NULL, palette = "Reds")
}
\arguments{
\item{map}{a data object (currently supported objects are matrices, data frames,
and spatial objects from the sp package of classes SpatialPoints,
SpatialPointsDataFrame, Polygon, Polygons, SpatialPolygons,
SpatialPolygonsDataFrame, Line, Lines, SpatialLines, and SpatialLinesDataFrame)}

\item{choices}{a character vector specify the column name of map data which being used as a color scale}

\item{popupcol}{a string specify the column name of map data which being used as a pop-up window}

\item{palette}{The colors or color function that values will be mapped to}
}
\value{
A HTML widget object, on which we can add graphics layers using %>% (see examples)
}
\description{
Create a Leaflet map widget with multiple map layers and control
}

