% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{tbc}
\alias{tbc}
\title{A dataset contains counts of newly developed patients of a disease according
to Korean administrative area}
\format{A data frame with 16 rows and 3 columns
\describe{
  \item{시도별}{Korean administrative area}
  \item{Y2001}{total number of newly developed patients in the area in 2001}
  \item{Y2002}{total number of newly developed patients in the area in 2002}
  \item{Y2003}{total number of newly developed patients in the area in 2003}
  \item{Y2004}{total number of newly developed patients in the area in 2004}
  \item{Y2005}{total number of newly developed patients in the area in 2005}
  \item{Y2006}{total number of newly developed patients in the area in 2006}
  \item{Y2007}{total number of newly developed patients in the area in 2007}
  \item{Y2008}{total number of newly developed patients in the area in 2008}
  \item{Y2009}{total number of newly developed patients in the area in 2009}
  \item{Y2010}{total number of newly developed patients in the area in 2010}
  \item{Y2011}{total number of newly developed patients in the area in 2011}
  \item{Y2012}{total number of newly developed patients in the area in 2012}
  \item{Y2013}{total number of newly developed patients in the area in 2013}
  \item{code}{code indicating Korean administrative area}
}}
\usage{
tbc
}
\description{
A dataset contains counts of newly developed patients of a disease according
to Korean administrative area
}
\keyword{datasets}

